requirements.txt
Flask==3.0.2
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
openai==1.35.14
notion-client==2.2.1
